requirements.txt
fastapi
uvicorn
langchain
langchain-openai
faiss-cpu
pypdf
python-multipart
openai
langchain-community
python-dotenv
tiktoken

